+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Add entry attachments API (`WriteAttachments`, `ReadAttachments`, `RemoveAttachments`) with integration tests for API v1.19, [PR-41](https://github.com/reductstore/reduct-go/pull/41)
+
 ## 1.18.0 - 2026-02-04
 
 ### Added